+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.1] - Unreleased
 
+### Added
+
+- Add alias for verify method as scan.
+
 ## [0.2.0] - 2025-02-26
 
 ### Changed